+++ src/Portfolio/ClientApp/app.tsx
@@ -1,0 +1,5 @@
+import { createRoot } from 'react-dom/client';
+
+import { AppComponent } from './components/app.component';
+
+createRoot(document.getElementById('react-root') as HTMLElement).render(<AppComponent />);