+++ src/Portfolio/ClientApp/components/home.component.tsx
@@ -1,0 +1,13 @@
+import { AppBody, AppHeader } from './ux';
+
+export function HomeComponent() {
+  return <>
+    <AppHeader>
+      <ol className="breadcrumb flex-fill border-start">
+        <li className="breadcrumb-item active">Home</li>
+        <div className="ms-auto"></div>
+      </ol>
+    </AppHeader>
+    <AppBody></AppBody>
+  </>
+}