+++ src/Portfolio/Common/HttpUtils.cs
@@ -1,0 +1,5 @@
+namespace Portfolio.Common;
+
+public static class HttpUtils {
+  public static string GetAuthToken(this HttpRequest request) => request.Headers["x-auth-token"].FirstOrDefault();
+}