+++ src/Portfolio/Core/Auth/UserService.cs
@@ -1,0 +1,11 @@
+using Portfolio.Common;
+
+namespace Portfolio.Core;
+
+public class UserService {
+  public bool IsAdmin(string username) {
+    if (username.IsEmpty()) return false;
+
+    return true;
+  }
+}