+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Next
 
+- Add `Login` screen
+- Add `Home` screen
+- User Management
+- Expense Management
+  - Add `Dashboard` screen
+  - Add `Search` screen
+  - Add `Add` screen
+  - Add `Edit` screen
+
+## 0.0.2
+> 2024-07-16
+
+- Add `Import` screen
+  - Extract daily reporting from excel file
+
 ## 0.0.1
 > 2024-07-07